bugfix: cpu-only build path fixed

+++ src/line_renderer.h
@@ -11,5 +11,5 @@ public:
 
 private:
   unsigned int VAO, VBO;
-  int lineCount;
+  size_t lineCount;
 };
